+++ pkg/frontend/transport/retry.go
@@ -78,12 +78,12 @@ func (r *Retry) Do(ctx context.Context, f func() (*httpgrpc.HTTPResponse, error)
 }
 
 func isBodyRetryable(body string) bool {
-	// If pool exhausted or concurrent bytes limit exceeded, retry at query frontend
+	// If pool exhausted or concurrent data bytes limit exceeded, retry at query frontend
 	// might make things worse. Rely on retries at querier level only.
 	if strings.Contains(body, pool.ErrPoolExhausted.Error()) {
 		return false
 	}
-	if strings.Contains(body, storegateway.ErrMaxConcurrentBytesLimitExceeded.Error()) {
+	if strings.Contains(body, storegateway.ErrMaxConcurrentDataBytesLimitExceeded.Error()) {
 		return false
 	}
 	return true